Add files via upload
Police flasher (strobe) circuit, a signal consisting of a pair of led flashing in sequence This project consists of a frequency generating 555 timer and a timer operating in astable mode. It counts the incoming pulses and is strengthened with BD135 according to the number of pulses at the outputs. The 10-counter that lights up red LEDs is built on the 40107 IC.